[ads] General housekeeping
