+++ components/brave_ads/core/internal/BUILD.gn
@@ -1240,16 +1240,16 @@ static_library("internal") {
     "user_engagement/conversions/types/verifiable_conversion/envelope/verifiable_conversion_envelope_util.cc",
     "user_engagement/conversions/types/verifiable_conversion/envelope/verifiable_conversion_envelope_util.h",
     "user_engagement/conversions/types/verifiable_conversion/envelope/verifiable_conversion_envelope_util_constants.h",
+    "user_engagement/conversions/types/verifiable_conversion/id_pattern/parsers/verifiable_conversion_id_html_meta_tag_parser_util.cc",
+    "user_engagement/conversions/types/verifiable_conversion/id_pattern/parsers/verifiable_conversion_id_html_meta_tag_parser_util.h",
+    "user_engagement/conversions/types/verifiable_conversion/id_pattern/parsers/verifiable_conversion_id_html_parser_util.cc",
+    "user_engagement/conversions/types/verifiable_conversion/id_pattern/parsers/verifiable_conversion_id_html_parser_util.h",
+    "user_engagement/conversions/types/verifiable_conversion/id_pattern/parsers/verifiable_conversion_id_url_redirects_parser_util.cc",
+    "user_engagement/conversions/types/verifiable_conversion/id_pattern/parsers/verifiable_conversion_id_url_redirects_parser_util.h",
+    "user_engagement/conversions/types/verifiable_conversion/id_pattern/verifiable_conversion_id_pattern_util.cc",
+    "user_engagement/conversions/types/verifiable_conversion/id_pattern/verifiable_conversion_id_pattern_util.h",
     "user_engagement/conversions/types/verifiable_conversion/verifiable_conversion_builder.cc",
     "user_engagement/conversions/types/verifiable_conversion/verifiable_conversion_builder.h",
-    "user_engagement/conversions/types/verifiable_conversion/verifiable_conversion_id_pattern/parsers/verifiable_conversion_id_html_meta_tag_parser_util.cc",
-    "user_engagement/conversions/types/verifiable_conversion/verifiable_conversion_id_pattern/parsers/verifiable_conversion_id_html_meta_tag_parser_util.h",
-    "user_engagement/conversions/types/verifiable_conversion/verifiable_conversion_id_pattern/parsers/verifiable_conversion_id_html_parser_util.cc",
-    "user_engagement/conversions/types/verifiable_conversion/verifiable_conversion_id_pattern/parsers/verifiable_conversion_id_html_parser_util.h",
-    "user_engagement/conversions/types/verifiable_conversion/verifiable_conversion_id_pattern/parsers/verifiable_conversion_id_url_redirects_parser_util.cc",
-    "user_engagement/conversions/types/verifiable_conversion/verifiable_conversion_id_pattern/parsers/verifiable_conversion_id_url_redirects_parser_util.h",
-    "user_engagement/conversions/types/verifiable_conversion/verifiable_conversion_id_pattern/verifiable_conversion_id_pattern_util.cc",
-    "user_engagement/conversions/types/verifiable_conversion/verifiable_conversion_id_pattern/verifiable_conversion_id_pattern_util.h",
     "user_engagement/conversions/types/verifiable_conversion/verifiable_conversion_info.cc",
     "user_engagement/conversions/types/verifiable_conversion/verifiable_conversion_info.h",
     "user_engagement/reactions/reactions.cc",
